+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.2.0] - 2018-03-06
+### Added
+- src/reporter.js, basic reporting, to be expanded
+
+### Changed
+- src/proto/prototype.creep.js, added changeWorkroom() method
+                                refactored findStructure() method, includes now Containers
+- src/proto/prototype.spawn.js, improved spawnCustomCreep() method
+- src/roles/role.builder.js, bug fix
+- src/roles/role.harvester.js, implemented changeWorkroom() for Harvesters
+                               Harvesters use extra energy for building Structures
+- src/main.js, implements Reporter every 50 ticks
+
 ## [3.1.0] - 2018-03-06
 ### Added
 - src/proto/prototype.creep.js, basic functionalities